xhci: Fix DCI connection breaks during XHCI reset
XHCI reset is causing temprorary DCI connection break which results in
IPC_Error during reset call. It seems that the issue might be silently
handled on Windows hosts, but takes place on Linux hosts.

Add a work-around to fix reset call: catch an exception and wait for 2
seconds to let DCI connection be restored.